Rename lua/jj to lua/telescope-jj
Fixed #6
