+++ lua/telescope-jj/init.lua
@@ -1,0 +1,9 @@
+local conflicts = require("telescope-jj.conflicts")
+local diff = require("telescope-jj.diff")
+local files = require("telescope-jj.files")
+
+return {
+    conflicts = conflicts,
+    diff = diff,
+    files = files,
+}
