+++ lua/telescope/_extensions/jj.lua
@@ -8,5 +8,5 @@ return telescope.register_extension({
     setup = function(ext_config, config)
         -- access extension config and user config
     end,
-    exports = require("jj"),
+    exports = require("telescope-jj"),
 })
